Split out saveEdited
